django_authen_api/requirements.txt
appdirs==1.4.4
asgiref==3.4.1
autopep8==1.5.7
backports.entry-points-selectable==1.1.0
black==21.7b0
cfgv==3.3.0
click==8.0.1
dataclasses==0.8
distlib==0.3.2
Django==3.2.5
djangorestframework==3.12.4
djangorestframework-simplejwt==4.4.0
filelock==3.0.12
flake8==3.9.2
identify==2.2.12
importlib-metadata==4.6.3
importlib-resources==5.2.2
isort==5.9.3
mccabe==0.6.1
mypy-extensions==0.4.3
nodeenv==1.6.0
pathspec==0.9.0
platformdirs==2.2.0
pre-commit==2.13.0
pycodestyle==2.7.0
pyflakes==2.3.1
PyJWT==1.7.1
pytz==2021.1
PyYAML==5.4.1
regex==2021.8.3
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
tomli==1.2.1
typed-ast==1.4.3
typing-extensions==3.10.0.0
virtualenv==20.7.0
zipp==3.5.0
